Use markdown instead of Latex for the resume
Latex is very painful to work with, especially with this template. So,
use markdown instead so it is much simpler to maintain the resume.

The latex template is taken (with modifications) from Luke Smith's
resume over at [0].

Future plans include moving my website over to using pandoc too with
this markdown file.

[0] https://github.com/LukeSmithxyz/md-website-cv